refactor: [#7 - 리뷰반영] 사용자 정보 호출 메소드 책임 변경
사용자의 로그인 요청 값으로 사용자 정보를 가져오는 책임을 UserService에서 담당하도록
변경하였습니다.

UserController에서 담당하는 사용자 정보 처리 부분을 SessionAuthService에서
UserService를 통해 가져오도록 변경하였습니다.
