+++ src/main/java/com/campool/service/AuthService.java
@@ -1,12 +1,10 @@
 package com.campool.service;
 
-import com.campool.model.UserSignUp;
+import com.campool.model.UserLoginRequest;
 
 public interface AuthService {
 
-    UserSignUp getByIdAndPw(String id, String password);
-
-    void authenticate(UserSignUp userSignUp);
+    void authenticate(UserLoginRequest userSignUp);
 
     boolean isValidAuthentication();
